+++ repositories/user_repository.go
@@ -18,12 +18,16 @@ func NewUserRepository(DB *gorm.DB) *UserRepository {
 	return &UserRepository{DB: DB}
 }
 
-// Create 插入新的用户记录
+// Create 插入新的用户记录， ID必须为0
 func (r *UserRepository) Create(user *models.User) error {
-	var existingUser models.User
+
+	// 判断ID是否符合规范
+	if user.ID != 0 {
+		return error_code.ErrInvalidID
+	}
 
 	// 检查用户名中是否符合规范
-	if !utils.IsValidUsername(user.Username) {
+	if !utils.IsValidName(user.Username) {
 		return error_code.ErrInvalidUsername
 	}
 
@@ -34,11 +38,11 @@ func (r *UserRepository) Create(user *models.User) error {
 
 	// 检查描述是否符合规范
 	if !utils.IsValidDescription(user.Description) {
-		return error_code.ErrInvalidInput
+		return error_code.ErrInvalidDescription
 	}
 
 	// 检查邮箱是否被占用
-	if err := r.DB.Where("email = ?", user.Email).First(&existingUser).Error; err == nil {
+	if err := r.DB.Where("email = ?", user.Email).First(&models.User{}).Error; err == nil {
 		return error_code.ErrEmailExists
 	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
 		// 捕获潜在系统错误
@@ -46,7 +50,7 @@ func (r *UserRepository) Create(user *models.User) error {
 	}
 
 	// 检查用户名是否已存在
-	if err := r.DB.Where("username = ?", user.Username).First(&existingUser).Error; err == nil {
+	if err := r.DB.Where("username = ?", user.Username).First(&models.User{}).Error; err == nil {
 		return error_code.ErrUsernameExists
 	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
 		// 捕获潜在系统错误
@@ -95,16 +99,15 @@ func (r *UserRepository) Read(ID uint, email, username string) (*models.User, er
 	return &user, nil
 }
 
-// Update 更新用户记录, 参数：*models.User{ID, ...}
+// Update 更新用户记录, ID必须存在
 func (r *UserRepository) Update(user *models.User) error {
-	var existingUser models.User
 	// 检查用户ID是否有效
 	if user.ID == 0 {
-		return error_code.ErrInvalidInput
+		return error_code.ErrInvalidID
 	}
 
 	// 检查用户ID是否存在
-	if err := r.DB.Where("id = ?", user.ID).First(&existingUser).Error; err != nil {
+	if err := r.DB.Where("id = ?", user.ID).First(&models.User{}).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
 			return error_code.ErrUserNotFound
 		}
@@ -112,7 +115,7 @@ func (r *UserRepository) Update(user *models.User) error {
 	}
 
 	// 检查用户名中是否符合规范
-	if !utils.IsValidUsername(user.Username) {
+	if !utils.IsValidName(user.Username) {
 		return error_code.ErrInvalidUsername
 	}
 
@@ -123,11 +126,11 @@ func (r *UserRepository) Update(user *models.User) error {
 
 	// 检查描述是否符合规范
 	if !utils.IsValidDescription(user.Description) {
-		return error_code.ErrInvalidInput
+		return error_code.ErrInvalidDescription
 	}
 
 	// 检查邮箱是否被占用
-	if err := r.DB.Where("email = ? AND id != ?", user.Email, user.ID).First(&existingUser).Error; err == nil {
+	if err := r.DB.Where("email = ? AND id != ?", user.Email, user.ID).First(&models.User{}).Error; err == nil {
 		return error_code.ErrEmailExists
 	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
 		// 捕获潜在系统错误
@@ -135,7 +138,7 @@ func (r *UserRepository) Update(user *models.User) error {
 	}
 
 	// 检查用户名是否已存在
-	if err := r.DB.Where("username = ? AND id != ?", user.Username, user.ID).First(&existingUser).Error; err == nil {
+	if err := r.DB.Where("username = ? AND id != ?", user.Username, user.ID).First(&models.User{}).Error; err == nil {
 		return error_code.ErrUsernameExists
 	} else if !errors.Is(err, gorm.ErrRecordNotFound) {
 		// 捕获潜在系统错误
@@ -149,9 +152,17 @@ func (r *UserRepository) Update(user *models.User) error {
 	return nil
 }
 
-// Delete 删除用户记录
+// Delete 删除用户记录 ID必须存在
 func (r *UserRepository) Delete(user *models.User) error {
+	// 判断用户ID是否有效
+	if user.ID == 0 {
+		return error_code.ErrInvalidID
+	}
+
 	if err := r.DB.Delete(user).Error; err != nil {
+		if errors.Is(err, gorm.ErrRecordNotFound) {
+			return error_code.ErrUserNotFound
+		}
 		// 捕获系统错误
 		return error_code.ErrInternalServer
 	}
